% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_name.R
\name{classify_name}
\alias{classify_name}
\title{Classify a Fungal Name}
\usage{
classify_name(name)
}
\arguments{
\item{name}{A character string representing a fungal name.}
}
\value{
A list with two elements:
\itemize{
\item \code{type}: A character string indicating the classification type ("gen." or "sp.").
\item \code{name}: The base name of the fungus (either genus or species).
}
}
\description{
This function classifies a given fungal name into either a genus or species.
It returns the type (\code{gen.} for genus, \code{sp.} for species) and the base name.
}
