% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean and Format Fungal Names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{A character string representing the fungal name to be cleaned.}
}
\value{
A character string representing the cleaned fungal name.
}
\description{
This function takes a fungal name and applies basic cleaning and formatting to it,
including extracting the genus and species (or genus-species combination) and removing unnecessary parts.
}
