% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_updated_id.R
\name{get_updated_id}
\alias{get_updated_id}
\title{Get the Most Recent ID of a Fungal Record}
\usage{
get_updated_id(df, name)
}
\arguments{
\item{df}{A tibble or dataframe containing fungal records.}

\item{name}{A character string representing the fungal species name.}
}
\value{
A character string representing the ID of the most recently updated record.
}
\description{
This function extracts the most recent ID for a given fungal species from a dataframe,
sorting the records by the update date and returning the ID of the most recent entry.
}
